+++ src/main/java/ru/yandex/practicum/filmorate/dal/mappers/UserRowMapper.java
@@ -13,7 +13,7 @@ public class UserRowMapper implements RowMapper<User> {
     @Override
     public User mapRow(ResultSet rs, int rowNum) throws SQLException {
         User user = new  User();
-        user.setId(rs.getLong("id"));
+        user.setId(rs.getInt("id"));
         user.setName(rs.getString("name"));
         user.setLogin(rs.getString("login"));
         user.setBirthday(Objects.requireNonNull(rs.getDate("birthday")).toLocalDate());
