+++ src/main/java/ru/yandex/practicum/filmorate/dal/storage/friendship/FriendshipDbStorage.java
@@ -15,14 +15,14 @@ public class FriendshipDbStorage implements FriendshipStorage {
     }
 
     @Override
-    public void addFriend(Long userId, Long friendId) {
+    public void addFriend(Integer userId, Integer friendId) {
         final String sql = "insert into friendships (user_id, friend_id) values (?, ?)";
 
         jdbc.update(sql, userId, friendId);
     }
 
     @Override
-    public void deleteFriend(Long userId, Long friendId) {
+    public void deleteFriend(Integer userId, Integer friendId) {
         final String sql = "delete from friendships where user_id = ? and friend_id = ?";
 
         jdbc.update(sql, userId, friendId);
