+++ src/main/java/ru/yandex/practicum/filmorate/dal/storage/friendship/FriendshipStorage.java
@@ -1,7 +1,7 @@
 package ru.yandex.practicum.filmorate.dal.storage.friendship;
 
 public interface FriendshipStorage {
-    void addFriend(Long userId, Long friendId);
+    void addFriend(Integer userId, Integer friendId);
 
-    void deleteFriend(Long userId, Long friendId);
+    void deleteFriend(Integer userId, Integer friendId);
 }
