+++ src/main/java/ru/yandex/practicum/filmorate/dal/storage/like/LikeStorage.java
@@ -2,7 +2,7 @@ package ru.yandex.practicum.filmorate.dal.storage.like;
 
 public interface LikeStorage {
 
-    void addLikeToFilm(Long filmId, Long userId);
+    void addLikeToFilm(Integer filmId, Integer userId);
 
-    void deleteLikeFromFilm(Long filmId, Long userId);
+    void deleteLikeFromFilm(Integer filmId, Integer userId);
 }
